worker/requirements.txt
# Core dependencies
openai>=1.40.0
elevenlabs>=1.2.0
moviepy==1.0.3
python-dotenv>=1.0
tqdm>=4.66
requests>=2.32
pydub>=0.25.1
supabase>=2.10.0
httpx>=0.27.0
imageio-ffmpeg>=0.4.9
mutagen>=1.47.0

# Whisper for word alignment
whisper-timestamped>=1.14.0

# Image processing
Pillow>=10.0.0

# For future API/SaaS
fastapi>=0.109.0
uvicorn>=0.27.0
celery>=5.3.0
redis>=5.0.0

# Database (for SaaS)
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Cloud storage
boto3>=1.34.0  # For S3/R2
